+++ src/Parser.php
@@ -35,12 +35,14 @@ class Parser
         'size' => 0,
         'start' => 0,
         'seek' => 0,
+        'chunkSize' => 0,
         // data pre-manipulation
         'autotrim' => true,
         'onBeforeColumnParse' => null,
         'guesser' => null,
         // data post-manipulation
         'onRowParsed' => null,
+        'onChunkParsed' => null,
         'onError' => null,
         // column definition
         'columns' => [],
@@ -103,7 +105,7 @@ class Parser
      */
     public function fromFile(string $file, array $options = []) : array
     {
-        $options['close'] = isset($options['close']) ? $options['close'] : true;
+        $options['close'] = true;
         return $this->parse(new CsvFile($file, $options), $options);
     }
 
@@ -116,7 +118,7 @@ class Parser
      */
     public function fromString(string $data, array $options = []) : array
     {
-        $options['close'] = isset($options['close']) ? $options['close'] : true;
+        $options['close'] = true;
         return $this->parse(new CsvString($data, $options), $options);
     }
 
@@ -213,6 +215,7 @@ class Parser
             : 0;
 
         $parsed = [];
+        $chunks = 0;
         while (
             (!$options['size'] || $i < $options['size'] + $options['start']) &&
             false !== ($row = $data->getCsv($options['length'], $options['delimiter'], $options['enclosure'], $options['escape']))
@@ -233,6 +236,15 @@ class Parser
             } catch (\Exception $e) {
                 $this->checkError($e, $index, [ 'type' => 'row' ], $options);
             }
+            if (
+                is_callable($options['onChunkParsed']) &&
+                $options['chunkSize'] &&
+                count($parsed) === $options['chunkSize']
+            ) {
+                $this->postProcess($parsed, $columns, $options, $chunks);
+                $chunks += 1;
+                $parsed = [];
+            }
             $i++;
         }
         if ($options['close']) {
@@ -241,7 +253,7 @@ class Parser
         if (!count($parsed)) {
             return $parsed;
         }
-        return $this->postProcess($parsed, $columns, $options);
+        return $this->postProcess($parsed, $columns, $options, $chunks);
     }
 
     /**
@@ -250,9 +262,10 @@ class Parser
      * @param array $data the processed data
      * @param array $columns columns metadata
      * @param array $options configuration options for parsing
+     * @param int $indexCall number of times this method was called
      * @return array the processed data
      */
-    private function postProcess(array $data, array $columns, array $options) : array
+    private function postProcess(array $data, array $columns, array $options, int $indexCall) : array
     {
         $keys = array_keys($data[0]);
         $result = [];
@@ -285,6 +298,13 @@ class Parser
                 } catch (\Exception $e) {
                     $this->checkError($e, $index, $meta, $options);
                 }
+            }
+        }
+        if ($options['chunkSize'] && is_callable($options['onChunkParsed'])) {
+            try {
+                $options['onChunkParsed']($data, $indexCall, $columns, $options);
+            } catch (\Exception $e) {
+                $this->checkError($e, $indexCall, [ 'type' => 'chunk' ], $options);
             }
         }
         return $data;
@@ -402,8 +422,7 @@ class Parser
             $h = preg_replace('/\s\s+/', ' ', str_replace(["\r\n", "\r", "\n"], ' ', $h));
             if (isset($headers[$h])) {
                 $e = new Exception(sprintf("Header [%s] can't be the same for two columns", $h), Exception::HEADEREXISTS);
-                $meta = [ 'type' => 'init', 'key' => $h ];
-                $this->checkError($e, null, $meta, $options);
+                $this->checkError($e, null, [ 'type' => 'init', 'key' => $h ], $options);
             }
             $headers[$h] = $i;
         }