+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 - added an algorithm to guess line ending and delimiter
 - breaking `CsvInterface` now implements 2 new methods: `getResource` and `setMetadata`
 
+## Release 4.2.1
+
+- small change in chunk parsing, so it can be catched if an exception occurs
+- docs
+
+## Release 4.2.0
+
+- added `chunkSize` and `onChunkParsed` options (see documentation)
+
 ## Release 4.1.0
 
 - added `fromSplFile()` method and `\voilab\csv\CsvSplFile` class to manage `\SplFileObject`